+++ app/src/main/java/com/openull/eastroots92/vacation_homework_android/presenter/Main/MainContract.java
@@ -1,4 +1,4 @@
-package com.openull.eastroots92.vacation_homework_android.ui.presenter;
+package com.openull.eastroots92.vacation_homework_android.presenter.Main;
 
 public interface MainContract {
     interface View{
