+++ app/src/main/java/com/openull/eastroots92/vacation_homework_android/presenter/Main/MainPresenter.java
@@ -1,4 +1,4 @@
-package com.openull.eastroots92.vacation_homework_android.ui.presenter;
+package com.openull.eastroots92.vacation_homework_android.presenter.Main;
 
 public class MainPresenter implements MainContract.Presenter {
     MainContract.View view;
